user permission

+++ main_app/views.py
@@ -1,3 +1,4 @@
+from distutils.log import Log
 from django.shortcuts import render, redirect
 from django.contrib.auth import login
 from django.contrib.auth.forms import UserCreationForm
@@ -5,6 +6,8 @@ from django.views.generic import ListView, DetailView
 from django.views.generic.edit import CreateView, UpdateView, DeleteView
 from django.db.models import Avg
 from django.contrib.auth.models import User
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.mixins import LoginRequiredMixin
 
 from .models import Restaurant, MenuItem, Recipe, Ingredient, Instruction, Review
 from .forms import InstructionForm, IngredientForm, ReviewForm
@@ -31,7 +34,7 @@ def signup(request):
         if form.is_valid():
             user = form.save()
             login(request, user)
-            return redirect('index')
+            return redirect('home')
         else:
             error_message = 'Invalid sign up - try again'
     form = UserCreationForm()
@@ -76,23 +79,32 @@ class RecipeDetail(DetailView):
         context['review_form'] = ReviewForm()
         return context
 
-class RecipeCreate(CreateView):
+
+class RecipeCreate(LoginRequiredMixin, CreateView):
     model = Recipe
     fields = '__all__'
     success_url= '/recipes/'
+    login_url = 'signup'
 
     def get_success_url(self) -> str:
         return f'/recipes/{self.object.id}'
 
-class RecipeUpdate(UpdateView):
+
+class RecipeUpdate(LoginRequiredMixin, UpdateView):
     model = Recipe
     fields = '__all__'
+    login_url = 'signup'
 
-class RecipeDelete(DeleteView):
+    def form_valid(self, form):
+        form.instance.user = self.request.user
+        return super().form_valid(form)
+
+class RecipeDelete(LoginRequiredMixin, DeleteView):
     model = Recipe
     success_url = '/'
+    login_url = 'signup'
 
-
+@login_required
 def add_instruction(request, recipe_id):
     form = InstructionForm(request.POST)
     if form.is_valid():
@@ -101,6 +113,7 @@ def add_instruction(request, recipe_id):
         new_instruction.save()
     return redirect('recipes_detail', pk=recipe_id)
 
+@login_required
 def add_ingredient(request, recipe_id):
     form = IngredientForm(request.POST)
     if form.is_valid():
@@ -109,18 +122,22 @@ def add_ingredient(request, recipe_id):
         new_ingredient.save()
     return redirect('recipes_detail', pk=recipe_id)
 
+@login_required
 def delete_ingredient(request, ingredient_id):
     ingredient = Ingredient.objects.get(id=ingredient_id)
     recipe = ingredient.recipe
-    ingredient.delete()
+    if request.user == ingredient.user:
+        ingredient.delete()
     return redirect('recipes_detail', pk=recipe.id)
 
+@login_required
 def delete_instruction(request, instruction_id):
     instruction = Instruction.objects.get(id=instruction_id)
     recipe = instruction.recipe
     instruction.delete()
     return redirect('recipes_detail', pk=recipe.id)
 
+@login_required
 def add_review(request, recipe_id):
     form = ReviewForm(request.POST)
     if form.is_valid():
@@ -130,6 +147,7 @@ def add_review(request, recipe_id):
         new_review.save()
     return redirect('recipes_detail', pk=recipe_id)
 
+@login_required
 def delete_review(request, review_id):
     review = Review.objects.get(id=review_id)
     recipe = review.recipe
